xfsdump: remove BMV_IF_NO_DMAPI_READ flag
Use of the flag has had no effect since kernel commit 288699fecaff
("xfs: drop dmapi hooks") which removed all dmapi related code, so we
can remove it.

Given that there are no other flags that need to be specified for the
bmap call, convert once instance of it from getbmapx to plain getbmap.

Signed-off-by: Anthony Iliopoulos <ailiop@suse.com>
Reviewed-by: Carlos Maiolino <cmaiolino@redhat.com>
Signed-off-by: Carlos Maiolino <cem@kernel.org>